110A got rejected for 4744000695826. Added long long

+++ codeforces/110A-Nearly Lucky Number/main.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main()
 {
-    long number;
+    long long number;
     cin >> number;
     int count = 0;
     while (number != 0)
